:boom: Regex and module rules combined
The previous 'module' rule no longer exists. The same functionality was duplicated in the 'regex' rule introduced in v1.2.1. Now, only the 'module' rule exists, but with full regex functionality.

Breaking change flagged due to:

1. Deletion of 'regex' rule
2. Change in API for the 'module' rule

Details on how to change existing usages will be included in the GitHub release notes.